requirements.txt
Flask>=2.0
pytesseract
Pillow
streamlit==1.26.1
pytesseract==0.3.11
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.26.0
tk==0.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
openpyxl